requirements.txt
ultralytics
flask
waitress
pillow
Flask_Cors
Werkzeug
pybase64